backend: Introduce flakeutils.py
